\name{bfastPlot-package}
\alias{bfastPlot-package}
\alias{bfastPlot}
\docType{package}
\title{
Plot bfast-type objects
}
\description{
Plot bfast-type objects...more to come....
}
\details{
\tabular{ll}{
Package: \tab bfastPlot\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-30\cr
License: \tab What license is it under?\cr
}
}
\author{
Ben DeVries <devries.br@gmail.com>
}
\references{

}

\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
