% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bfmPlot}
\alias{bfmPlot}
\title{ggplot bfastmonitor}
\usage{
bfmPlot(bfm, plotlabs = NULL, ncols = 1, rescale = 1, ylab = "response",
  displayMagn = TRUE, magn_ypos = 0.3, magn_xoffset = -0.45,
  magn_digits = 3, displayTrend = TRUE, displayResiduals = c("none",
  "all", "monperiod", "history"), type = "irregular")
}
\description{
Generate a ggplot object from a (list of) bfastmonitor object(s)
}
\examples{
# adapted from help page of bfastmonitor

library(bfast)
library(ggplot2)

NDVIa <- as.ts(zoo(som$NDVI.a, som$Time))
plot(NDVIa)
## apply the bfast monitor function on the data
## start of the monitoring period is c(2010, 13)
## and the ROC method is used as a method to automatically identify a stable history
mona1 <- bfastmonitor(NDVIa, start = c(2010, 13), formula = response ~ harmon, order = 3)
class(mona1)

# regular plot
plot(mona1)
# ggplot of the same
p <- bfmPlot(mona1)
p

## the advantage of ggplot is that is is object based
## additional layers can simply be added
# change to black/white
p2 <- p + theme_bw()
p2

## combine several bfastmonitor objects into one facet plot
mona2 <- bfastmonitor(NDVIa, start = c(2010, 13), formula = response~harmon, order=2)
mona3 <- bfastmonitor(NDVIa, start = c(2010, 13), formula = response~harmon, order=1)
p3 <- bfmPlot(list(mona1, mona2, mona3), plotlabs = c("order = 3", "order = 2", "order = 1"))
p3 + theme_bw()

# it's not necessary to show the trend when there is none
p4 <- bfmPlot(list(mona1, mona2, mona3),
plotlabs = c("order = 3", "order = 2", "order = 1"),
displayTrend = FALSE)
p4 + theme_bw()
}

