% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bfmPredict}
\alias{bfmPredict}
\title{Harmonic and trend predictions from a bfastmonitor object}
\usage{
bfmPredict(bfm, type = "irregular", plotlabs = NULL)
}
\arguments{
\item{bfm}{Object (or list of objects) of type \code{bfastmonitor}}

\item{type}{Character. Type of time series on which \code{bfm} is based. Either ``irregular" or ``16-day" (see \code{\link{bfastts}}).}

\item{plotlabs}{Character. Vector of labels corresponding to objects in \code{bfm}. Can be omitted (automatic labels will be added in that case if \code{bfm} is a list of bfastmonitor objects.)}
}
\value{
\code{data.frame} with model parameters and predictions according to bfastmontior parameters in \code{bfm}. See details for more info.
}
\description{
Produce a new data.frame from bfastmonitor objects including harmonic and trend predictions (where applicable). Multiple bfastmonitor objects can be used (see details).
}

