% Generated by roxygen2 (4.0.2): do not edit by hand
\name{time2date}
\alias{time2date}
\title{Convert decimal date to yyyy-mm-dd}
\usage{
time2date(time)
}
\arguments{
\item{time}{Numeric. Time in decimal format (e.g. 2005.258)}
}
\value{
Object of class \code{Date}
}
\description{
Convert decimal date to yyyy-mm-dd
}
\examples{
time2date(2005.259)
}

